backend/pyproject.toml
[tool.poetry]
name = "bridgecard-status-page-backend"
version = "0.1.0"
description = ""
authors = ["enfinity <festusowumi@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
firebase-admin = "5.0.1"
python-dotenv = "0.19.0"
pydantic = "1.8.2"
fastapi = "0.67.0"
fastapi-versioning = "0.9.1"
dependency-injector = "4.41.0"
uvicorn = "0.14.0"
aiofiles = "0.7.0"
sse-starlette = "^1.6.5"
anyio = "3.3.0"
sh = "^2.0.6"
pymongo = "^4.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
